refactored multiple files and moved logic to relevant files.
