+++ resources/quotes.js
